+++ test/RayTest.cpp
@@ -12,8 +12,8 @@ namespace {
 		protected:
 			void SetUp() {
 
-				r.o = Vector2f(1,1);
-				r.d = Vector2f(1,1.7320);
+				r.o = Vector2d(1,1);
+				r.d = Vector2d(1,1.7320);
 				r.frequency = 2.5e6;
 			}
 
@@ -25,12 +25,12 @@ namespace {
 		ASSERT_NEAR(1.047, r.getAngle(), 0.001);
 
 		Ray r2 = Ray();
-		r2.d = Vector2f(1, -1.732);
+		r2.d = Vector2d(1, -1.732);
 
 		ASSERT_NEAR(-1.047, r2.getAngle(), 0.001);
 
 		Ray r3 = Ray();
-		r3.d = Vector2f(-10, 10);
+		r3.d = Vector2d(-10, 10);
 
 		ASSERT_NEAR(2.356, r3.getAngle(), 0.001);
 	}
